possible auto leave fix
