+++ src/main/kotlin/command/PlayCommand.kt
@@ -1,8 +1,11 @@
 package command
 
 import GlobalData
-import audio.DefaultAudioLoadResultHandler
 import audio.GuildAudioManager
+import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler
+import com.sedmelluq.discord.lavaplayer.tools.FriendlyException
+import com.sedmelluq.discord.lavaplayer.track.AudioPlaylist
+import com.sedmelluq.discord.lavaplayer.track.AudioTrack
 import discord4j.core.event.domain.message.MessageCreateEvent
 import reactor.core.publisher.Mono
 
@@ -19,7 +22,7 @@ class PlayCommand : Command {
     private fun play(audioManager: GuildAudioManager, event: MessageCreateEvent) {
         val query = event.message.content.substringAfter(" ")
         val track = getTrack(query)
-        GlobalData.PLAYER_MANAGER.loadItem(track, DefaultAudioLoadResultHandler(audioManager, event))
+        GlobalData.PLAYER_MANAGER.loadItem(track, defaultAudioLoadResultHandler(audioManager, event))
     }
 
     private fun getTrack(query: String): String {
@@ -27,6 +30,37 @@ class PlayCommand : Command {
             return query
         }
         return GlobalData.SEARCH_CLIENT.getTracksForSearch(query)[0].url
+    }
+
+    private fun defaultAudioLoadResultHandler(
+        audioManager: GuildAudioManager,
+        event: MessageCreateEvent
+    ): AudioLoadResultHandler {
+        return object : AudioLoadResultHandler {
+            override fun trackLoaded(track: AudioTrack) {
+                println("trackLoaded")
+                audioManager.scheduler.play(track)
+                event.message.channel
+                    .flatMap { it.createMessage("Track added: ${track.info.title}") }
+                    .block()
+            }
+
+            override fun playlistLoaded(playlist: AudioPlaylist) {
+                println("playlistLoaded")
+                playlist.tracks.forEach { audioManager.scheduler.play(it) }
+                event.message.channel
+                    .flatMap { it.createMessage("Playlist ${playlist.name} added with ${playlist.tracks.size} tracks") }
+                    .block()
+            }
+
+            override fun noMatches() {
+                println("no matches")
+            }
+
+            override fun loadFailed(exception: FriendlyException?) {
+                println("load failed")
+            }
+        }
     }
 
 }
